+++ src/1-one/contraction-one/model/index.js
@@ -1,2 +1,4 @@
 import contractions from './contractions.js'
-export default { one: { contractions } }
+import units from './units.js'
+
+export default { one: { contractions, units } }